+++ src/core/event_loop.rs
@@ -75,6 +75,15 @@ fn main_task<B: Backend + std::io::Write>(
         }
     };
 
+    // Check if there is a current song playing on startup
+    if let Some((_, song)) = ctx.find_current_song_in_queue()
+        && let Some(command) = &ctx.config.on_song_change {
+            let mut env = create_env(&ctx, std::iter::empty());
+            
+            run_external(command.clone(), env);
+        }
+    }
+
     match ctx.status.state {
         State::Play => {
             // Start update loop since a song is playing on startup